for higher port counts, these fields must be extended to 64 bit
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@61398 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-6.6/include/linux/phy.h
@@ -417,10 +417,10 @@ struct mii_bus {
 	struct mdio_device *mdio_map[PHY_MAX_ADDR];
 
 	/** @phy_mask: PHY addresses to be ignored when probing */
-	u32 phy_mask;
+	u64 phy_mask;
 
 	/** @phy_ignore_ta_mask: PHY addresses to ignore the TA/read failure */
-	u32 phy_ignore_ta_mask;
+	u64 phy_ignore_ta_mask;
 
 	/**
 	 * @irq: An array of interrupts, each PHY's interrupt at the index
